Line Chart + Stock Chart Addition
Added line charts for various stock tickers. This keeps track of the stock price changes in real time.

Quick problem to address: we are using AlphaVantage for this project which has a free tier but unfortunately has a delay. Therefore the line chart is not instantaneous. This is something t keep in mind for now but I will continue with this.